+++ CHANGELOG.md
@@ -65,7 +65,6 @@
 - 'normalise' flags was added to the plot_remap_frame routine, to normalize the plot to unity
 - **bug**: Solved a bug in the get frame index
 - **bug**: Solved a small bug, the .cin files was not being closed in the read_settings structure
-- Export remap: temporally not available, until it is adapted to the new remap structure
 - Translate remap: temporally not available, until it is adapted to the new remap structure
 - getTimeTrace now returns also the used mask
 - exp_dat, remap_dat are now xarrays instead of dictionaries
@@ -85,6 +84,8 @@
 - getGyroradius and getEnergy now uses the proper amu to kg conversion and no longer rely on the proton mass
 - Integrate_remap change completely inputs and output, now it can handle also any translation of the remap. Please have a look at the new function doc. Take care with the translation if you do not use standard remaps as initial point
 - The small bug on the units of the weights of INPA markers from SINPA was corrected (the correction due to the detector pinhole size was already performed in SINPA)
+- Included new sub-library MHD, to calculate/plot basic mode parameters such as mode frequencies 
+
 
 ## 0.9.10: iHIBP videos improvements and TORBEAM sims.
 - Corrected bugs in the iHIBPsim videos read.
@@ -93,7 +94,7 @@
 
 ## 0.9.9: BEP removed from library and work in iHIBP.
 - Removed the full BEP library from the ScintSuite.
-- Rebranded class for the iHIBP video with:
+- Re-branded class for the iHIBP video with:
     - Distorted and aligned scintillator.
     - Plotting time traces of the signal within the scintillator.
     - Basic and advanced methods to substract noise from the signal.